+++ miso-web/src/main/webapp/scripts/form_project.js
@@ -141,9 +141,51 @@ FormTarget.project = (function($) {
           sortSource: Utils.sorting.standardSort('alias'),
           required: true,
           nullLabel: '(Choose)'
+        }, {
+          title: 'Samples Expected',
+          data: 'samplesExpected',
+          type: 'int',
+          min: 1
+        }, {
+          title: 'Contact',
+          type: 'special',
+          makeControls: function(form) {
+            var label = object.contactId ? Contacts.makeContactLabel(object.contactName, object.contactEmail) : 'n/a';
+            return [$('<span>').attr('id', 'contactLabel').text(label), makeButton('Change', function() {
+              showContactDialog(object, form);
+            }), makeButton('Remove', function() {
+              setContact(null, object, form);
+            }), $('<div>').attr('id', 'projectForm_contactError').addClass('errorContainer')];
+          }
         }]
       }];
     }
   };
+
+  function makeButton(text, callback) {
+    return $('<button>').addClass('ui-state-default').attr('type', 'button').css('margin-left', '4px').text(text).click(callback);
+  }
+
+  function showContactDialog(project, form) {
+    Contacts.selectContactDialog(false, false, function(contact) {
+      setContact(contact, project, form);
+    });
+  }
+
+  function setContact(contact, project, form) {
+    $('#projectForm_contactError').empty();
+    if (contact) {
+      project.contactId = contact.id;
+      project.contactName = contact.name;
+      project.contactEmail = contact.email;
+      $('#contactLabel').text(Contacts.makeContactLabel(contact.name, contact.email));
+    } else {
+      project.contactId = null;
+      project.contactName = null;
+      project.contactEmail = null;
+      $('#contactLabel').text('n/a');
+    }
+    form.markOtherChanges();
+  }
 
 })(jQuery);